Moved HtmlInfo class into its own files.
