+++ src/gui/InfoPanel.cpp
@@ -15,10 +15,12 @@
     You should have received a copy of the GNU General Public License
     along with Ready. If not, see <http://www.gnu.org/licenses/>.         */
 
-// local:
 #include "InfoPanel.hpp"
+
+// local:
 #include "app.hpp"              // for wxGetApp
 #include "frame.hpp"
+#include "HtmlInfo.hpp"
 #include "IDs.hpp"
 #include "prefs.hpp"            // for readydir, etc
 #include "wxutils.hpp"          // for Warning, CopyTextToClipboard
@@ -30,7 +32,6 @@
 
 // wxWidgets:
 #include <wx/filename.h>        // for wxFileName
-#include <wx/html/htmlwin.h>    // for wxHtmlWindow
 #include <wx/colordlg.h>
 
 // VTK:
@@ -41,272 +42,31 @@
 #include <string>
 using namespace std;
 
-const wxString change_prefix = _("change: ");
-
-// labels in 1st column
-const wxString rule_name_label = _("Rule name");
-const wxString rule_type_label = _("Rule type");
-const wxString description_label = _("Description");
-const wxString num_chemicals_label = _("Number of chemicals");
-const wxString formula_label = _("Formula");
-const wxString kernel_label = _("Kernel");
-const wxString dimensions_label = _("Dimensions");
-const wxString block_size_label = _("Block size");
-const wxString number_of_cells_label = _("Number of cells");
-const wxString wrap_label = _("Toroidal wrap-around");
-const wxString data_type_label = _("Data type");
-const wxString neighborhood_type_label = _("Neighborhood");
-const wxString neighborhood_range_label = _("Neighborhood range");
-const wxString neighborhood_weight_label = _("Neighborhood weight");
-
 // -----------------------------------------------------------------------------
 
-// define a child window for displaying HTML info
-class HtmlInfo : public wxHtmlWindow
-{
-    public:
-
-        HtmlInfo(wxWindow* parent, MyFrame* myframe, wxWindowID id = wxID_ANY,
-            const wxPoint& pos = wxDefaultPosition, const wxSize& size = wxDefaultSize,
-            long style = wxHW_SCROLLBAR_AUTO | wxWANTS_CHARS)
-            : wxHtmlWindow(parent, id, pos, size, style)
-        {
-            frame = myframe;
-            panel = (InfoPanel*)parent;
-            editlink = false;
-            linkrect = wxRect(0,0,0,0);
-            scroll_x = scroll_y = 0;
-        }
-
-        virtual void OnLinkClicked(const wxHtmlLinkInfo& link);
-        virtual void OnCellMouseHover(wxHtmlCell* cell, wxCoord x, wxCoord y);
-
-        void ClearStatus();  // clear pane's status line
-
-        void SetFontSizes(int size);
-        void ChangeFontSizes(int size);
-
-        wxRect linkrect;     // rect for cell containing link
-
-        void SaveScrollPos() { GetViewStart(&scroll_x, &scroll_y); }
-        void RestoreScrollPos() { Scroll(scroll_x, scroll_y); }
-        void ResetScrollPos() { Scroll(0,0); }
-
-    private:
-
-        void OnSize(wxSizeEvent& event);
-        void OnMouseMotion(wxMouseEvent& event);
-        void OnMouseLeave(wxMouseEvent& event);
-        void OnMouseDown(wxMouseEvent& event);
-        void OnHtmlCellClicked(wxHtmlCellEvent& event);
-
-    private:
-
-        MyFrame* frame;
-        InfoPanel* panel;
-
-        bool editlink;          // open clicked file in editor?
-        int scroll_x,scroll_y;  // remember the scroll position
-
-        DECLARE_EVENT_TABLE()
-};
-
-BEGIN_EVENT_TABLE(HtmlInfo, wxHtmlWindow)
-    EVT_SIZE          (HtmlInfo::OnSize)
-    EVT_MOTION        (HtmlInfo::OnMouseMotion)
-    EVT_ENTER_WINDOW  (HtmlInfo::OnMouseMotion)
-    EVT_LEAVE_WINDOW  (HtmlInfo::OnMouseLeave)
-    EVT_LEFT_DOWN     (HtmlInfo::OnMouseDown)
-    EVT_RIGHT_DOWN    (HtmlInfo::OnMouseDown)
-    EVT_HTML_CELL_CLICKED (wxID_ANY, HtmlInfo::OnHtmlCellClicked)
-END_EVENT_TABLE()
+const wxString InfoPanel::rule_name_label = _("Rule name");
+const wxString InfoPanel::rule_type_label = _("Rule type");
+const wxString InfoPanel::description_label = _("Description");
+const wxString InfoPanel::num_chemicals_label = _("Number of chemicals");
+const wxString InfoPanel::formula_label = _("Formula");
+const wxString InfoPanel::kernel_label = _("Kernel");
+const wxString InfoPanel::dimensions_label = _("Dimensions");
+const wxString InfoPanel::block_size_label = _("Block size");
+const wxString InfoPanel::number_of_cells_label = _("Number of cells");
+const wxString InfoPanel::wrap_label = _("Toroidal wrap-around");
+const wxString InfoPanel::data_type_label = _("Data type");
+const wxString InfoPanel::neighborhood_type_label = _("Neighborhood");
+const wxString InfoPanel::neighborhood_range_label = _("Neighborhood range");
+const wxString InfoPanel::neighborhood_weight_label = _("Neighborhood weight");
 
 // -----------------------------------------------------------------------------
-
-void HtmlInfo::OnLinkClicked(const wxHtmlLinkInfo& link)
-{
-    wxString url = link.GetHref();
-    if (url.StartsWith(wxT("http:")) || url.StartsWith(wxT("https:")) || url.StartsWith(wxT("mailto:"))) {
-        // pass http/mailto URL to user's preferred browser/emailer
-        if ( !wxLaunchDefaultBrowser(url) )
-            Warning(_("Could not open URL in browser!"));
-
-    } else if ( url.StartsWith(change_prefix) ) {
-        panel->ChangeInfo( url.Mid(change_prefix.size()) );
-        // best to reset focus after dialog closes
-        SetFocus();
-
-    } else if ( url.StartsWith(wxT("prefs:")) ) {
-        // user clicked on link to Preferences dialog
-        frame->ShowPrefsDialog( url.AfterFirst(':') );
-
-    } else if ( url.StartsWith(wxT("edit:")) ) {
-        // open clicked file in user's preferred text editor
-        wxString path = url.AfterFirst(':');
-        #ifdef __WXMSW__
-            path.Replace(wxT("/"), wxT("\\"));
-        #endif
-        wxFileName fname(path);
-        if (!fname.IsAbsolute()) path = readydir + path;
-        frame->EditFile(path);
-
-
-    } else if ( url.StartsWith(wxT("open:")) ) {
-        // open clicked file
-        wxString path = url.AfterFirst(':');
-        #ifdef __WXMSW__
-            path.Replace(wxT("/"), wxT("\\"));
-        #endif
-        wxFileName fname(path);
-        if (!fname.IsAbsolute()) path = readydir + path;
-        if (editlink) {
-            frame->EditFile(path);
-        } else {
-            frame->Raise();
-            frame->OpenFile(path);
-        }
-
-    } else {
-        // assume it's a link to a local target
-        LoadPage(url);
-    }
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::OnHtmlCellClicked(wxHtmlCellEvent& event)
-{
-    wxHtmlCell* cell = event.GetCell();
-    int x = event.GetPoint().x;
-    int y = event.GetPoint().y;
-
-    wxHtmlLinkInfo* link = cell->GetLink(x,y);
-    if (link) {
-        // set linkrect to avoid bug in wxHTML if click was in outer edge of link
-        // (bug is in htmlwin.cpp in wxHtmlWindowMouseHelper::HandleIdle;
-        // OnCellMouseHover needs to be called if cell != m_tmpLastCell)
-        wxPoint pt = ScreenToClient( wxGetMousePosition() );
-        linkrect = wxRect(pt.x-x, pt.y-y, cell->GetWidth(), cell->GetHeight());
-    }
-
-    event.Skip();   // call OnLinkClicked
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::OnCellMouseHover(wxHtmlCell* cell, wxCoord x, wxCoord y)
-{
-    wxHtmlLinkInfo* link = cell->GetLink(x,y);
-    if (link) {
-        wxString href = link->GetHref();
-        href.Replace(wxT("&"), wxT("&&"));
-        panel->SetStatus(href);
-        wxPoint pt = ScreenToClient( wxGetMousePosition() );
-        linkrect = wxRect(pt.x-x, pt.y-y, cell->GetWidth(), cell->GetHeight());
-    } else {
-        ClearStatus();
-    }
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::OnMouseMotion(wxMouseEvent& event)
-{
-    if (!linkrect.IsEmpty()) {
-        int x = event.GetX();
-        int y = event.GetY();
-        if (!linkrect.Contains(x,y)) ClearStatus();
-    }
-    event.Skip();
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::OnMouseLeave(wxMouseEvent& event)
-{
-    ClearStatus();
-    event.Skip();
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::ClearStatus()
-{
-    panel->SetStatus(wxEmptyString);
-    linkrect = wxRect(0,0,0,0);
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::OnMouseDown(wxMouseEvent& event)
-{
-    // set flag so ctrl/right-clicked file can be opened in editor
-    editlink = event.ControlDown() || event.RightDown();
-    event.Skip();
-}
-
-// -----------------------------------------------------------------------------
-
-// avoid scroll position being reset to top when wxHtmlWindow is resized
-void HtmlInfo::OnSize(wxSizeEvent& event)
-{
-    SaveScrollPos();
-
-    Freeze(); // prevent flicker
-
-    wxHtmlWindow::OnSize(event);
-
-    wxString currpage = GetOpenedPage();
-    if ( !currpage.IsEmpty() ) {
-        LoadPage(currpage);         // reload page
-    }
-
-    RestoreScrollPos();
-    Thaw();
-
-    // prevent wxHtmlWindow::OnSize being called again
-    event.Skip(false);
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::SetFontSizes(int size)
-{
-    // set font sizes for <FONT SIZE=-2> to <FONT SIZE=+4>
-    int f_sizes[7];
-    f_sizes[0] = int(size * 0.6);
-    f_sizes[1] = int(size * 0.8);
-    f_sizes[2] = size;
-    f_sizes[3] = int(size * 1.2);
-    f_sizes[4] = int(size * 1.4);
-    f_sizes[5] = int(size * 1.6);
-    f_sizes[6] = int(size * 1.8);
-    #ifdef __WXOSX_COCOA__
-        SetFonts(wxT("Lucida Grande"), wxT("Monaco"), f_sizes);
-    #else
-        SetFonts(wxEmptyString, wxEmptyString, f_sizes);
-    #endif
-}
-
-// -----------------------------------------------------------------------------
-
-void HtmlInfo::ChangeFontSizes(int size)
-{
-    // changing font sizes resets pos to top, so save and restore pos
-    SaveScrollPos();
-    SetFontSizes(size);
-    RestoreScrollPos();
-    panel->UpdateButtons();
-}
-
-// -----------------------------------------------------------------------------
-
-// define a panel that contains a HtmlInfo window
 
 BEGIN_EVENT_TABLE(InfoPanel, wxPanel)
     EVT_BUTTON (ID::SmallerButton,  InfoPanel::OnSmallerButton)
     EVT_BUTTON (ID::BiggerButton,   InfoPanel::OnBiggerButton)
 END_EVENT_TABLE()
+
+// -----------------------------------------------------------------------------
 
 InfoPanel::InfoPanel(MyFrame* parent, wxWindowID id)
     : wxPanel(parent,id), frame(parent)
@@ -510,7 +270,7 @@ wxString InfoPanel::AppendRow(const wxString& print_label, const wxString& label
 
     if (is_editable) {
         result += _T("<td valign=top align=right><a href=\"");
-        result += change_prefix;
+        result += HtmlInfo::change_prefix;
         result += label;
         result += _T("\">");
         result += _("edit");
